# Avoid allocating in Keys methods